+++ egui_commonmark_backend/src/pulldown.rs
@@ -185,4 +185,5 @@ pub fn parser_options() -> Options {
         | Options::ENABLE_TASKLISTS
         | Options::ENABLE_STRIKETHROUGH
         | Options::ENABLE_FOOTNOTES
+        | Options::ENABLE_DEFINITION_LIST
 }